Minimize the performance impact of unsafe handles if they aren't used
This commit adapts the implementation of unsafe handles, so that unless
they are used, the performance impact is only two empty checks against
a synchronized map in `FiberContext.evaluateNow`. On the JVM, an
additional optimization makes sure, that the performance impact
increases linearly with the number of unsafe handles, and not linearly
with the number of `FiberRef`s. This optimisation is not available
for JS, since it is based on a `WeakHashMap`.