Add support for setting fiber refs from side effecting code
This commit completely overhauls the old implementation, that was based
on a single `ThreadLocal` for every `FiberRef` with an `UnsafeHandle`.
The new implementation uses only one, global `ThreadLocal` to store the
currently executing `FiberContext`. This not only greatly simplifies the
code for `unsafeGet`, it also makes implementing `unsafeSet` easily
possible.