Don't loose updates when refs/handles are used from multiple fibers
This change introduces a second, thread local, namely
`FiberRef.pendingUnsafeSets`, to make sure that updates aren't lost
even when `FiberContext.current` is not yet set. As a side effect,
the semantics of `FiberRef.UnsafeHandle` are that of an ordinary
thread local if used after the termination of the ZIO interpreter.